+++ discord/src/main/kotlin/lt/saltyjuice/dragas/chatty/v3/discord/message/general/Field.kt
@@ -2,8 +2,15 @@ package lt.saltyjuice.dragas.chatty.v3.discord.message.general
 
 import com.google.gson.annotations.SerializedName
 
-open class Field
+open class Field()
 {
+    @JvmOverloads
+    constructor(name: String, value: String, inline: Boolean = false) : this()
+    {
+        this.name = name
+        this.value = value
+        this.inline = inline
+    }
     @SerializedName("name")
     var name: String = ""